bpo-26317: scproxy is an Obj-C File
This file includes an Objective-C header from macOS, which makes the
whole translation unit an Objective-C file.

AppleClang is relaxed about this but GCC needs an additional hint to
understand that this is not a standard C file.

This fixes silent (!) `_scproxy` GCC build issues on macOS
```
                  from /Users/runner/runners/2.169.1/work/cpython/cpython/Modules/_scproxy.c:6:
/Library/Developer/CommandLineTools/SDKs/MacOSX10.15.sdk/System/Library/Frameworks/Security.framework/Headers/Authorization.h:193:7: error: variably modified 'bytes' at file scope
  193 |  char bytes[kAuthorizationExternalFormLength];
      |       ^~~~~
```

that corrupt the installed python interpreter on usage.